Adding a copy of the LXX2012
Added brackets around inserted words, removed a bunch of the XML, played
around a little bit with the formatting. There is still some XML that
needs to be removed, but I think the majority of it is between books so
hopefully it won't get in the way.  Also updated the README with
additional info.